pyproject.toml
[tool.poetry]
name = "bitcoind-zmq-checker"
version = "0.1.0"
description = ""
authors = ["Kenichi Sato <ksato9700@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pyzmq = "^19.0.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
